status codes and error handling

+++ Controllers/SongsController.cs
@@ -20,35 +20,51 @@ namespace MusicRestAPI.Controllers
             _dbContext = dbContext;
         }
         [HttpGet]
-        public IEnumerable<Song> GetAllSongs()
+        public IActionResult GetAllSongs()
         {
-            return _dbContext.Songs.ToList();
+            return Ok(_dbContext.Songs.ToList());
         }
         [HttpGet("{id}")]
-        public Song GetSongById(int id)
+        public IActionResult GetSongById(int id)
         {
-            return _dbContext.Songs.Find(id);
+            var existing = _dbContext.Songs.Find(id);
+            if (existing == null)
+            {
+                return NotFound("No record found against this id");
+            }
+            return Ok(existing);
         }
         [HttpPost]
-        public void AddSong([FromBody]Song song)
+        public IActionResult AddSong([FromBody]Song song)
         {
             _dbContext.Songs.Add(song);
             _dbContext.SaveChanges();
+            return StatusCode(StatusCodes.Status201Created);
         }
         [HttpPut("{id}")]
-        public void UpdateSong(int id, [FromBody]Song song)
+        public IActionResult UpdateSong(int id, [FromBody]Song song)
         {
             var existing = _dbContext.Songs.Find(id);
+            if (existing==null)
+            {
+                return NotFound("No record found against this id");
+            }
             existing.Title = song.Title;
             existing.Language = song.Language;
             _dbContext.SaveChanges();
+            return Ok("Record updated successfully");
         }
         [HttpDelete("{id}")]
-        public void DeleteSong(int id)
+        public IActionResult DeleteSong(int id)
         {
             var existing = _dbContext.Songs.Find(id);
+            if (existing == null)
+            {
+                return NotFound("No record found against this id");
+            }
             _dbContext.Songs.Remove(existing);
             _dbContext.SaveChanges();
+            return Ok("Record deleted successfully");
         }
     }
 }
